fix(review): coordinate initialization link recovery
